requirements.txt
tensorflow==2.18.0
flask
pillow
numpy
gunicorn
gdown
